Moves the editor into the main object route, and uses the correct isManifest check; removes typeStore since nothing is using it
