--- services/admin/src/routes/object/[prefix]/edit/[noid].svelte
@@ -1,70 +1,0 @@
-<script lang="ts">
-  import type { AccessObject } from "@crkn-rcdr/access-data";
-  import { isCanvasManifest, isCollection } from "@crkn-rcdr/access-data";
-  import ManifestContentEditor from "$lib/components/manifests/ManifestContentEditor.svelte";
-  import SideMenuPageListButton from "$lib/components/shared/SideMenuPageListButton.svelte";
-  import SideMenuPage from "$lib/components/shared/SideMenuPage.svelte";
-  import EditorLayout from "$lib/components/access-objects/EditorLayout.svelte";
-  import { getContext } from "svelte";
-
-  let object: AccessObject;
-  const objectStore: SvelteStore<AccessObject> = getContext("objectStore");
-  $: object = $objectStore;
-  $: setDataModel(object);
-
-  let type: "collection" | "canvasManifest" | "other" = "other"; // ts complains with lack of default
-  const typeStore: SvelteStore<"collection" | "canvasManifest" | "other"> =
-    getContext("typeStore");
-  $: type = $typeStore;
-
-  let rfdc: any; // Deep copies an object
-  let model: AccessObject;
-  async function setDataModel(object: AccessObject) {
-    rfdc = (await import("rfdc")).default();
-    model = rfdc(object) as AccessObject;
-  }
-</script>
-
-<slot />
-
-{#if object && model}
-  <!-- I couldn't have named slots in a regular __layout, not sure if there's a smarter way to make the layout more extensible-->
-  <!-- I get an error if the check is in the EditorLayout about the slots-->
-  {#if isCanvasManifest(model)}
-    <EditorLayout bind:object bind:model>
-      <div slot="editor-menu">
-        <SideMenuPageListButton>Content</SideMenuPageListButton>
-      </div>
-
-      <div slot="editor-content">
-        <SideMenuPage overflowY="hidden">
-          <ManifestContentEditor bind:manifest={model} />
-        </SideMenuPage>
-      </div>
-    </EditorLayout>
-  {:else if isCollection(model)}
-    <EditorLayout bind:object bind:model>
-      <div slot="editor-menu">
-        <!-- SideMenuPageListButton>Content</SideMenuPageListButton-->
-      </div>
-
-      <div slot="editor-content">
-        <!--SideMenuPage overflowY="hidden">
-        Collection Content! {JSON.stringify(object)}
-      </SideMenuPage-->
-      </div>
-    </EditorLayout>
-  {:else}
-    <EditorLayout bind:object bind:model>
-      <div slot="editor-menu">
-        <!-- SideMenuPageListButton>Content</SideMenuPageListButton-->
-      </div>
-
-      <div slot="editor-content">
-        <!--SideMenuPage overflowY="hidden">
-        Other Content! {JSON.stringify(object)}
-      </SideMenuPage-->
-      </div>
-    </EditorLayout>
-  {/if}
-{/if}
